+++ String/KMP/KMP2.cpp
