--- String/Z Algorithm for finding longest prefix presented as substring.cpp
@@ -1,57 +1,0 @@
-#pragma warning(disable:4786)
-#pragma warning(disable:4996)
-#include<bits/stdc++.h>
-#define FOR(i, a, b) for (int i=a; i<(b); i++)
-#define F0R(i, a) for (int i=0; i<(a); i++)
-#define pii pair<int,int>
-#define pll pair<long long ,long long>
-#define pli pair<long long , int>
-#define pil pair<int ,long long>
-#define pi acos(-1)
-#define pb push_back
-#define mkp make_pair
-#define IOS ios::sync_with_stdio(0); cin.tie(0); cout.tie(0);
-#define ll long long
-#define MAX 300005
-#define INF 0x3f3f3f3f
-template <class T> inline T bigmod(T p,T e,T M){ll ret = 1LL;for(; e > 0LL; e >>= 1LL){if(e & 1LL) ret = (ret * p) % M;p = (p * p) % M;}return (T)ret;}
-template <class T> inline T modinverse(T a,T M){return bigmod(a,M-2,M);}   // M is prime}
-using namespace std;
-
-string Zfunction(string s){
-    int n=s.length();
-    int z[n];
-    int L = 0, R = 0;
-    for (int i = 1; i < n; i++) {
-      if (i > R) {
-        L = R = i;
-        while (R < n && s[R-L] == s[R]) R++;
-        z[i] = R-L; R--;
-      } else {
-        int k = i-L;
-        if (z[k] < R-i+1) z[i] = z[k];
-        else {
-          L = i;
-          while (R < n && s[R-L] == s[R]) R++;
-          z[i] = R-L; R--;
-        }
-      }
-    }
-    int ans;
-    int mx=-1;
-    for(int i=1;i<n;i++){
-        if(z[i]>mx){
-            ans=i;
-            mx=z[i];
-        }
-    }
-    string str(s.begin(),s.begin()+z[ans]);
-    return str;
-}
-int main(){
-	IOS
-	//freopen("output.txt","w",stdout);
-	string s;
-	cin>>s;
-    cout<<Zfunction(s);
-}
